refactor: split guest-agent integration tests
